not working model edit function

+++ app/src/main/java/com/example/waitr/HelperShape.kt
@@ -2,10 +2,10 @@ package com.example.waitr
 
 data class HelperShape(
     val id: String = "",
-    var height: Double = 0.0,
-    var width: Double = 0.0,
-    var xPosition: Double = 0.0,
-    var yPosition: Double = 0.0
+    var height: Int = 0,
+    var width: Int = 0,
+    var xPosition: Int = 0,
+    var yPosition: Int = 0
 ){
     fun toMap(): Map<String, Any?> {
         return mapOf(
