+++ app/src/main/java/com/example/waitr/Table.kt
@@ -6,11 +6,11 @@ data class Table(
     var state: String ="",
     var numberOfPeople: Int = 0,
     var listOfOrders: MutableList<Order> = mutableListOf(),
-    var totalOrderPrice: Double = 0.0,
-    var height: Double = 0.0,
-    var width: Double = 0.0,
-    var xPosition: Double = 0.0,
-    var yPosition: Double = 0.0
+    var totalOrderPrice: Int = 0,
+    var height: Int = 0,
+    var width: Int = 0,
+    var xPosition: Int = 0,
+    var yPosition: Int = 0
 ){
     fun toMap(): Map<String, Any?> {
         return mapOf(
